# backend/requests/openapi.yaml
openapi: 3.1.0
info:
  title: Instant Wellness Kits API
  description: |
    REST API for order management with automatic tax calculation and geocoding.

    **Key Features:**
    - 📍 Geocoding with Mapbox (New York State only)
    - 💰 Automatic tax calculation (state, county, city, MCTD)
    - 📊 Async CSV batch import with real-time progress
    - 📡 SSE (Server-Sent Events) for live updates
    - 📈 Detailed error reporting and job history

    **Tech Stack:**
    - Spring Boot 4.0.3
    - Spring Batch 6.0
    - PostgreSQL + Redis
    - Mapbox Geocoding API
  contact:
    name: Acheron Team
    email: support@acheron.com
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Local development server
- url: https://api.acheron.com
  description: Production server
...
paths:
  /tax-localities:
    get:
      tags:
      - Tax Localities
      summary: Get all tax localities
      description: "Returns all NY State tax localities with rates, reporting codes,\
        \ and district flags. Cached for 24h."
      operationId: getAllTaxLocalities
      responses:
        "200":
          description: List of tax localities
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxLocality"
    post:
      tags:
      - Tax Localities
      summary: Create a single tax locality
      operationId: createTaxLocality
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxLocality"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaxLocality"
  /tax-localities/reload:
    post:
      tags:
      - Tax Localities
      summary: Reload tax localities from built-in JSON
      description: Deletes ALL existing tax localities and reloads from the bundled
        ny-tax-localities.json file. Clears all caches.
      operationId: reloadFromJson
      responses:
        "200":
          description: Reloaded successfully
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
        "500":
          description: Failed to reload
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /tax-localities/import:
    post:
      tags:
      - Tax Localities
      summary: Import tax localities from custom JSON body
      description: Deletes ALL existing tax localities and replaces them with the
        provided JSON array. Clears all caches.
      operationId: importFromJson
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Imported successfully
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
        "400":
          description: Invalid JSON
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /orders:
    get:
      tags:
      - Orders
      summary: Get all orders
      description: Retrieves a paginated list of orders sorted by total amount
      operationId: getAllOrders
      parameters:
      - name: pageable
        in: query
        description: "Pagination parameters (page, size, sort)"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
    post:
      tags:
      - Orders
      summary: Create new order
      description: Creates a new order with automatic tax calculation based on coordinates
        using native GeoJSON point-in-polygon lookup. No external API calls.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "201":
          description: Order created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid request or location outside NY State
...
  /api/v1/native/import:
    post:
      tags:
      - Native Import
      summary: Import orders from CSV (native)
      description: "High-performance CSV import using virtual threads and native GeoJSON\
        \ point-in-polygon tax calculation. No external API calls. Processes ~11k\
        \ records/sec. CSV columns: id, longitude, latitude, timestamp, subtotal"
      operationId: importCsv
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file with columns: id, longitude, latitude, timestamp,\
                    \ subtotal"
              required:
              - file
      responses:
        "200":
          description: Import completed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Invalid CSV file
...
  /tax-localities/{locality}:
    get:
      tags:
      - Tax Localities
      summary: Find tax locality by name
      description: "Lookup a specific locality by name (case-insensitive, normalizes\
        \ county suffixes and NYC borough aliases)."
      operationId: getTaxLocality
      parameters:
      - name: locality
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Tax locality found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaxLocality"
        "404":
          description: Locality not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaxLocality"
    delete:
      tags:
      - Tax Localities
      summary: Delete a tax locality by name
      operationId: deleteTaxLocality
      parameters:
      - name: locality
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
...
components:
  schemas:
    TaxLocality:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        locality:
          type: string
        taxRatePercent:
          type: number
        reportingCode:
          type: string
        seeReference:
          type: string
        isSpecialDistrict:
          type: boolean
        isMctdDistrict:
          type: boolean
        parentCounty:
          type: string
        notes:
          type: string
    OrderRequest:
      type: object
      description: "Order details with latitude, longitude, and subtotal"
      properties:
        latitude:
          type: number
          maximum: 90.0
          minimum: -90.0
        longitude:
          type: number
          maximum: 180.0
          minimum: -180.0
        subtotal:
          type: number
      required:
      - latitude
      - longitude
      - subtotal
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        latitude:
          type: number
        longitude:
          type: number
        subtotal:
          type: number
        composite_tax_rate:
          type: number
        tax_amount:
          type: number
        total_amount:
          type: number
        timestamp:
          type: string
          format: date-time
        taxBreakdown:
          $ref: "#/components/schemas/TaxBreakdown"
        jurisdictions:
          type: array
          items:
            type: string
    TaxBreakdown:
      type: object
      properties:
        state_rate:
          type: number
        county_rate:
          type: number
        city_rate:
          type: number
        special_rates:
          type: number
    ImportError:
      type: object
      properties:
        lineNumber:
          type: integer
          format: int32
        errorMessage:
          type: string
        recordData:
          type: string
    ImportResult:
      type: object
      properties:
        totalRecords:
          type: integer
          format: int32
        successfulRecords:
          type: integer
          format: int32
        failedRecords:
          type: integer
          format: int32
        durationMs:
          type: integer
          format: int64
        recordsPerSecond:
          type: number
          format: double
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...